transform ApartmentAdapter to anonymous class
